restructure type system into focused files by purpose and usage
