+++ src/features/user/LoginForm.tsx
@@ -1,6 +1,6 @@
 import { MouseEvent, useState } from "react";
 import { useNavigate } from "react-router-dom";
-import { useAppDispatch } from "../../types/hooks";
+import { useAppDispatch } from "../../types/redux.hooks";
 import { setUserCredentials } from "../../store/slices/userSlice";
 import { validateUserCredentials } from "../../utils/helpers";
