+++ src/types/action.types.ts
@@ -1,0 +1,42 @@
+//All types related to action functions, including submission results and outcomes
+
+import { FeedbackFormErrors } from "./form.types";
+
+export type ActionType = "createFeedback" | "editFeedback" | "addComment";
+
+export type ValidationErrorResult = {
+  actionType: ActionType;
+  validationErrors: FeedbackFormErrors;
+};
+export type FailureResult = {
+  actionType: ActionType;
+  submitError: unknown;
+};
+export type SuccessResult<T> = {
+  actionType: ActionType;
+  payload: T;
+};
+
+/*Return value of createActionResult  */
+export type ActionResult<TPayload = null> =
+  | {
+      submissionOutcome: "validationError";
+      actionType: ActionType;
+      validationErrors: FeedbackFormErrors;
+      submitError: null;
+      payload: null;
+    }
+  | {
+      submissionOutcome: "failure";
+      actionType: ActionType;
+      validationErrors: null;
+      submitError: unknown;
+      payload: null;
+    }
+  | {
+      submissionOutcome: "success";
+      actionType: ActionType;
+      validationErrors: null;
+      submitError: null;
+      payload: TPayload;
+    };
